refactor: unify inference latency log type
